编译告警解决（Loss类算子）



# message auto-generated for no-merge-commit merge:
!81 merge br_warning_dep into master

编译告警解决（Loss类算子）

Created-by: init__zhb__
Commit-by: zhanghaobo
Merged-by: turing_project1
Description: ## 描述
编译Warning抑制 (Loss类算子)
| Warning 类型 | Warning详情 | 文件:行数 | 
| --- | --- |  ----- |
| `type-limits` |  warning: comparison of unsigned expression < 0 is always false | <ul> <li>`loss/ctc_loss_v3/op_host/ctc_loss_v3_tiling.cpp:405`</li> <li>`loss/ctc_loss_v3_grad/op_host/ctc_loss_v3_grad_tiling.cpp:377`</li></ul> |
| unused-parameter |   warning: unused parameter xxx | <ul> <li>`loss/binary_cross_entropy/op_host/op_api/aclnn_binary_cross_entropy.cpp:83`</li>  <li>`loss/binary_cross_entropy_grad/op_host/op_api/aclnn_binary_cross_entropy_backward.cpp:122`</li>  <li>`loss/nll_loss/op_host/op_api/aclnn_nll_loss.cpp:173`</li>  <li>`loss/nll_loss/op_host/op_api/aclnn_nll_loss.cpp:173`</li> <li>`loss/nll_loss/op_host/op_api/aclnn_nll_loss2d.cpp:165`</li> <li>`loss/nll_loss_grad/op_host/op_api/aclnn_nll_loss_backward.cpp:170`</li>  <li>`loss/sigmoid_cross_entropy_with_logits_grad_v2/op_host/op_api/aclnn_binary_cross_entropy_with_logits_backward.cpp:159,160,247 `</li> <li>`loss/sigmoid_cross_entropy_with_logits_v2/op_host/op_api/aclnn_binary_cross_entropy_with_logits.cpp:102,217`</li> <li>`loss/smooth_l1_loss_v2/op_host/op_api/aclnn_smooth_l1_loss.cpp:110`</li>  <li>`loss/soft_margin_loss/op_host/op_api/aclnn_soft_margin_loss.cpp:49:107: warning: unused parameter ‘reduction’`</li> <li>`loss/soft_margin_loss_grad/op_host/op_api/aclnn_soft_margin_loss_backward.cpp:93:47: warning: unused parameter ‘reduction’`</li> </ul> | 
| `unused-function` |  warning: xxx defined but not used | <ul> <li>`loss/common/level2_base_loss.h:28`</li> <li>`loss/common/level2_base_loss.h:44`</li></ul> |
| `sign-compare` | warning: comparison between signed and unsigned integer expressions | `loss/multilabel_margin_loss/op_host/op_api/aclnn_multilabel_margin_loss.cpp:87,96,97` | 

## 关联的Issue



## 测试
全量编译，没有上述文件的编译Warning信息
```bash
bash build.sh --make_clean_all
bash build.sh --pkg --soc=ascend910b 2> build_log_warnings_only.log
```

## 文档更新


## 类型标签
<!--  [x] 表示选中 -->
- [ ] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [x] 其他，请描述：[Warning]编译告警




See merge request: cann/ops-nn!81

+++ loss/ctc_loss_v3/op_host/ctc_loss_v3_tiling.cpp
@@ -409,7 +409,7 @@ ge::graphStatus TilingPrepare4CTCLossV3(gert::TilingParseContext* context)
         return ge::GRAPH_FAILED);
     compileInfo->sysWorkspaceSize = ascendcPlatform.GetLibApiWorkSpaceSize();
     OP_CHECK_IF(
-        (compileInfo->sysWorkspaceSize < 0),
+        (static_cast<uint64_t>(compileInfo->sysWorkspaceSize) < 0),
         OP_LOGE(nodeName, "sysWorkspaceSize should be greater than or equal to zero"),
         return ge::GRAPH_FAILED);
     uint64_t ubSizePlatForm;
